move class into shared and init in browser scripts
